Cleaned up strict compiler warnings in ide module

+++ ide/core/OtTaskManager.h
@@ -67,7 +67,8 @@ public:
 	size_t getRedoCount() { return redoStack.size(); }
 
 	// mark a new baseline
-	void baseline() { baselineMarker = int(undoStack.size()); }
+	void baseline() { baselineMarker = undoStack.size(); }
+
 	// see if document is "dirty" (i.e. it's off the saved baseline)
 	bool isDirty() { return undoStack.size() != baselineMarker; }
 
@@ -80,5 +81,5 @@ private:
 	std::chrono::time_point<std::chrono::system_clock> last;
 
 	// currently basedlined version (to support "dirty" tracking)
-	int baselineMarker = 0;
+	size_t baselineMarker = 0;
 };
